ZKC-31 reorg mpc/zkproofs into separate src files
